w3c-reports/12092014/efl/runned-tests/cors/status-preflight-expected.txt

Status after preflight


PASS CORS - status after preflight on GET 200 
PASS CORS - status after preflight on GET 204 
PASS CORS - status after preflight on GET 400 
PASS CORS - status after preflight on GET 401 
PASS CORS - status after preflight on HEAD 200 
PASS CORS - status after preflight on HEAD 204 
PASS CORS - status after preflight on HEAD 400 
PASS CORS - status after preflight on HEAD 401 
PASS CORS - status after preflight on HEAD 501 
FAIL CORS - status after preflight on HEAD 699 assert_equals: response status expected 699 but got 0(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/cors/status-preflight.htm:29:26
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
PASS CORS - status after preflight on POST 204 
PASS CORS - status after preflight on POST 400 
PASS CORS - status after preflight on POST 401 
PASS CORS - status after preflight on POST 404 
FAIL CORS - status after preflight on PUT 699 assert_equals: response status expected 699 but got 0(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/cors/status-preflight.htm:29:26
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
PASS CORS - status after preflight on CHICKEN 501 
